project/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bleach==6.1.0
crispy-bootstrap5==0.7
Django==4.2.6
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-crispy-forms==2.1
django-filter==23.3
django-summernote==0.8.20.0
django-taggit==4.0.0
gunicorn==21.2.0
packaging==23.2
Pillow==10.0.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
webencodings==0.5.1
whitenoise==6.6.0
